+++ CHANGELOG.md
@@ -5,7 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [1.1.0] / 2024-11-14
-- Update `ricaun.Revit.UI` to `0.7.*`
+- Update `ricaun.Revit.UI` to `0.7.1`
 - Add missing `Orange` color
 - Obsolete `RibbonImageThemeSelector` to use `ricaun.Revit.UI`.
 - Rename Images to have `Light` in the name to enable the theme change to `Dark`.